Refactor: temptopic visualizations
